requirements.txt
click==7.1.2
configparser==4.0.2
contextlib2==0.6.0.post1
Flask==1.1.4
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
importlib-metadata==2.1.3
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pathlib2==2.3.7.post1
scandir==1.10.0
six==1.16.0
SQLAlchemy==1.4.44
typing~=3.7.4.3
Werkzeug==1.0.1
zipp~=3.10.0

pip~=22.3.1
attr~=0.3.2
wheel~=0.37.1
colorama~=0.4.6
blinker~=1.5
greenlet~=2.0.1
setuptools~=65.5.1
opencv-python~=4.6.0.66
tensorflow~=2.10.0
numpy~=1.23.5

pytesseract~=0.3.10
Pillow~=9.3.0